finetune_qwen_classification/requirements.txt
# Qwen2.5-1.5B 文本分类微调依赖
# Python >= 3.8

# 核心框架
torch>=2.0.0
transformers>=4.35.0
modelscope>=1.11.0

# ms-swift 微调框架
ms-swift>=2.0.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# 通用工具
tqdm>=4.64.0
pyyaml>=6.0

# 可视化
matplotlib>=3.5.0
tensorboard>=2.11.0
